fix: experiments were not being saved correctly

+++ src/attacks/BadNets.py
@@ -4,6 +4,7 @@ import torch
 import numpy as np
 import matplotlib.pyplot as plt
 import csv
+import os
 
 
 class BadNets(Attack):
@@ -54,9 +55,10 @@ class BadNets(Attack):
                   'pos', 'color', 'seed', 'train_acc', 'train_loss', 'clean_acc',
                   'bk_acc', 'clean_loss', 'bk_loss']
 
-        with open(path_csv, 'w') as f:
-            writer = csv.writer(f)
-            writer.writerow(header)
+        if not os.path.exists(path_csv):
+            with open(path_csv, 'w') as f:
+                writer = csv.writer(f)
+                writer.writerow(header)
 
         with open(path_csv, 'a') as f:
             writer = csv.writer(f)
